+++ services/web/server/src/simcore_service_webserver/exporter/_handlers.py
@@ -15,6 +15,7 @@ from .._meta import API_VTAG
 from ..login.decorators import login_required
 from ..projects.api import with_project_locked_and_notify
 from ..projects.projects_api import retrieve_and_notify_project_locked_state
+from ..redis import get_redis_lock_manager_client_sdk
 from ..security.decorators import permission_required
 from ..users.api import get_user_fullname
 from ._formatter.archive import get_sds_archive_path
@@ -46,7 +47,7 @@ async def export_project(request: web.Request):
     assert project_uuid  # nosec
 
     @with_project_locked_and_notify(
-        request.app,
+        get_redis_lock_manager_client_sdk(request.app),
         project_uuid=project_uuid,
         status=ProjectStatus.EXPORTING,
         owner=Owner(
